+++ Sources/NablaCore/Utils/UmbrellaCancellable.swift
@@ -1,19 +1,24 @@
 import Foundation
-import NablaCore
 
-class UmbrellaCancellable: Cancellable {
-    func add(_ cancellable: Cancellable) {
+public class UmbrellaCancellable: Cancellable {
+    // MARK: - Public
+    
+    public func add(_ cancellable: Cancellable) {
         children.append(cancellable)
     }
     
-    private(set) var isCancelled: Bool = false
+    public private(set) var isCancelled: Bool = false
     
-    // MARK: - Cancellable
+    // MARK: Cancellable
     
-    func cancel() {
+    public func cancel() {
         children.removeAll() // Whe do not own the `children`, we only release them. Someone else might still use them.
         isCancelled = true
     }
+    
+    // MARK: Init
+    
+    public init() {}
     
     deinit {
         cancel()