+++ Sources/NablaMessagingCore/Domain/Interactors/CreateConversation/CreateConversationInteractor.swift
@@ -5,6 +5,7 @@ protocol CreateConversationInteractor {
     func execute(
         title: String?,
         providerIds: [UUID]?,
+        initialMessage: MessageInput?,
         handler: ResultHandler<Conversation, NablaError>
     ) -> Cancellable
 }